# Accelerate multimethod-calls with Vectorcall protocol