--- a/ci/azure-run-clang-format.yml
+++ b/ci/azure-run-clang-format.yml
@@ -7,6 +7,7 @@ steps:
   displayName: Install clang-format
 - bash: |
     clang-format-8 -i uarray/_uarray_dispatch.cxx uarray/small_dynamic_array.h
+    clang-format-8 -i uarray/vectorcall.cxx uarray/vectorcall.h
     if [ "$(git diff)" != "" ]; then
       echo "Detected formatting errors, suggested changes below:"
       git diff