--- a/uarray/vectorcall.cxx
+++ b/uarray/vectorcall.cxx
@@ -35,14 +35,110 @@ static PyObject * build_kwarg_dict(
   }
   return dict;
 }
-#elif ((PY_VERSION_HEX >= 0x03070000) && (PY_VERSION_HEX < 0x03090000))
-#  ifdef __cplusplus
-extern "C" {
-#  endif
-extern int _PyObject_GetMethod(PyObject *, PyObject *, PyObject **);
-#  ifdef __cplusplus
+#elif (PY_VERSION_HEX < 0x03090000)
+
+static int is_method_descr(PyTypeObject* descr_tp) {
+    return (
+        (descr_tp->tp_flags & Q_Py_TPFLAGS_METHOD_DESCRIPTOR) != 0 ||
+        (descr_tp == &PyFunction_Type) ||
+        (descr_tp == &PyMethodDescr_Type));
 }
-#  endif
+
+/* The below code is derivative of CPython source, and is taken because
+_PyObject_GetMethod is not exported from shared objects.
+
+Copyright (c) 2001, 2002, 2003, 2004, 2005, 2006, 2007, 2008, 2009, 2010,
+2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021 Python Software Foundation;
+All Rights Reserved
+*/
+
+/* Specialized version of _PyObject_GenericGetAttrWithDict
+   specifically for the LOAD_METHOD opcode.
+
+   Return 1 if a method is found, 0 if it's a regular attribute
+   from __dict__ or something returned by using a descriptor
+   protocol.
+
+   `method` will point to the resolved attribute or NULL.  In the
+   latter case, an error will be set.
+*/
+static int _PyObject_GetMethod(PyObject *obj, PyObject *name, PyObject **method)
+{
+    PyTypeObject *tp = Py_TYPE(obj);
+    PyObject *descr;
+    descrgetfunc f = NULL;
+    PyObject **dictptr, *dict;
+    PyObject *attr;
+    int meth_found = 0;
+
+    assert(*method == NULL);
+
+    if (Py_TYPE(obj)->tp_getattro != PyObject_GenericGetAttr
+            || !PyUnicode_Check(name)) {
+        *method = PyObject_GetAttr(obj, name);
+        return 0;
+    }
+
+    if (tp->tp_dict == NULL && PyType_Ready(tp) < 0)
+        return 0;
+
+    descr = _PyType_Lookup(tp, name);
+    if (descr != NULL) {
+        Py_INCREF(descr);
+        if (is_method_descr(Py_TYPE(descr))) {
+            meth_found = 1;
+        } else {
+            f = Py_TYPE(descr)->tp_descr_get;
+            if (f != NULL && PyDescr_IsData(descr)) {
+                *method = f(descr, obj, (PyObject *)Py_TYPE(obj));
+                Py_DECREF(descr);
+                return 0;
+            }
+        }
+    }
+
+    dictptr = _PyObject_GetDictPtr(obj);
+    if (dictptr != NULL && (dict = *dictptr) != NULL) {
+        Py_INCREF(dict);
+        attr = PyDict_GetItemWithError(dict, name);
+        if (attr != NULL) {
+            Py_INCREF(attr);
+            *method = attr;
+            Py_DECREF(dict);
+            Py_XDECREF(descr);
+            return 0;
+        }
+        else {
+            Py_DECREF(dict);
+            if (PyErr_Occurred()) {
+                Py_XDECREF(descr);
+                return 0;
+            }
+        }
+    }
+
+    if (meth_found) {
+        *method = descr;
+        return 1;
+    }
+
+    if (f != NULL) {
+        *method = f(descr, obj, (PyObject *)Py_TYPE(obj));
+        Py_DECREF(descr);
+        return 0;
+    }
+
+    if (descr != NULL) {
+        *method = descr;
+        return 0;
+    }
+
+    PyErr_Format(PyExc_AttributeError,
+                 "'%.50s' object has no attribute '%U'",
+                 tp->tp_name, name);
+    return 0;
+}
+
 #endif /* PYPY_VERSION */
 
 
@@ -101,7 +197,7 @@ PyObject * Q_PyObject_VectorcallDict(
 PyObject * Q_PyObject_VectorcallMethod(
     PyObject * name, PyObject * const * args, size_t nargsf,
     PyObject * kwnames) {
-#if (defined(PYPY_VERSION) || (PY_VERSION_HEX < 0x03070000))
+#ifdef PYPY_VERSION
   PyObject * callable = PyObject_GetAttr(args[0], name);
   if (!callable) {
     return NULL;
@@ -112,7 +208,7 @@ PyObject * Q_PyObject_VectorcallMethod(
   return result;
 #elif (PY_VERSION_HEX >= 0x03090000)
   return PyObject_VectorcallMethod(name, args, nargsf, kwnames);
-#else /* (PY_VERSION_HEX >= 0x03070000) */
+#else
   /* Private CPython code for CALL_METHOD opcode */
   PyObject * callable = NULL;
   int unbound = _PyObject_GetMethod(args[0], name, &callable);