@hameerabbasi 2021-04-20
Does SciPy still support Python 3.6? [NEP 29](https://numpy.org/neps/nep-0029-deprecation_policy.html#support-table) seems to say it doesn't. I guess we can drop support for it.