@peterbell10 2021-04-20
There's no code here that's specifically to support Python 3.6, so I think it's fine to leave support in for now.