@peterbell10 2021-04-20
> Would it be desirable to use the RAII wrappers in vectorcall.c, to avoid potential memory leaks?

I actually wrote it in plain C because I think this might be useful elsewhere. Each function only manages at most one object, so manual refcount isn't the _worst_ here.